refactor: address style debt, component bloat, and unused code
