+++ src/components/WidgetContainer/WidgetContainer.module.scss
@@ -5,6 +5,10 @@
   pointer-events: none;
 }
 
+.hidden {
+  visibility: hidden;
+}
+
 .dragWrapper {
   width: 100%;
   height: 100%;
